update build config
I am having trouble getting zig to find SDL2 in the buildPhase for nix,
so that is currently broken. It *does* work when I add it to an impure
shell and then run `zig build`, so that is something at least.

Not sure what the issue is with the nix builder. SDL *is* available,
`cc` can find <SDL2/SDL.h>, but for some reason zig cannot. Hopefully
we can fix this in future.