fix nix build
After experimenting with various command line options that zig accepts,
we now have a working build script for nix.

Zig seems to have some issue with linking correctly in our pure nix
build environment. Perhaps it is something to do with the `zig` input
to this flake, or we need to tell zig what ABI to target? If you remove
the patchelf command and examine the build artefact with
`ldd ./result/bin/shogi`, you'll see it has a line like this which
seems problematic:

/lib/ld-musl-x86_64.so.1 =>
    /nix/store/...-glibc-2.xx-xx/lib64/ld-linux-x86-64.so.2 (0x00...)

I'm not sure what is the best way to fix this, but for now we can
simply patch up the resulting binary and it works.