render pieces onto the board
We implement `fn renderPieces` which takes the current board state
and renders the pieces for each player. This gives us a basic starting
point for the rendering logic.

Once again, to accomplish this we set up more wrappers around the raw C
functions. This could be avoided, but not without adding a lot of
clutter and noise to the core logic. For readability when using
functions with many arguments, we use the struct trick to give each
argument a name and (optionally) a default value.

We also remove some `pub`s where they are needless.