initial pass at adding textures for every piece
It has proven non-trivial to organize these textures/images in a proper
data structure. So for now we simply use a dumb brute force approach
with a lot of repetition, in order to make progress.

As part of the effort to make this easier, we have re-organized our
core types to one flat enum. This will hopefully it easier to improve
the texture handling in future. As part of this effort, we remove the
now-unneeded `size` method, and add a `promote` method we will need
in future.