rename PlayerPiece to Piece
The names PlayerPiece and Piece, and variables like player_piece, are
not very clear to read. We can improve this significantly by simply
renaming them to something better, so that's exactly what we do in this
commit. We also rename local variables and struct fields of these types
to match.