rename Move -> Motion, add new Move type
The code would be more readable and self-explanatory if we had a proper
Move, one which remembers it's source and destination. So we rename
the current type to Motion to distinguish it's related but distinct
meaning, and create a new Move type.

Given this change, we then refactor the codebase to reflect and take
advantage of these types.