flatten module hierarchy
For ease of browsing and organising the codebase, we have flattened out
the module hierarchy. This required moving a few test cases and methods
around, and shifting some code to new modules `state.zig`
and `pixel.zig`. Notably we moved RenderError into `sdl.zig` and
renamed it to SdlError, and could then re-use it as needed
in `render.zig`.