add initial CPU prototype
Implements basic `cpu.zig` logic - just picks a random move for now as
a placeholder. Then we hook this up to the main game loop - after the
player makes a move, we spawn a separate thread in which the CPU's
move is calculated. Once this has completed, it stores that move in a
MutexGuard, and each loop we check that guard for pending moves to
apply to the game state.