refactor & only queue CPU move if user has moved
We need to return a bool from applyUserMove to indicate whether we
succeeded or not, in order to determine whether to call `queueCpuMove`.

We also adjust the MouseUp / MouseDown events to always set
the `.mouse.move_from` field, and separately handle the matter of which
mouse button was used to complete the move.