fix memory leaks & switch to HashMaps
There has been a persistent memory leak related to `rules.types.Valid`.
In the previous commit, it could be observed by simply running
`zig test source/test.zig`, or by running the program compiled in
Debug or ReleaseSafe modes and making at-least one move.

I have not been able to identify the exact cause of this leak. However,
it seems to no longer exist as a result of the changes in this commit.
At the same time, this is a more natural encoding to work with anyway,
using AutoHashMap to model possible moves. So, we will go ahead and
make this change.

Summary of changes:

- Switch type of `rules.types.Valid.basics`
  from `ArrayList( struct{ BoardPos, ArrayList(Movement) } )`
  to `AutoHashMap(BoardPos, ArrayList(Movement)`.
- Switch type of `rules.types.Valid.drops`
  from `ArrayList( struct{ Piece, ArrayList(BoardPos) } )`
  to `AutoHashMap(Piece, ArrayList(BoardPos)`.
- Add convenience init and deinit functions to these new 'basics'
  and 'drops' types.
- Refactor as needed to accommodate these new types.

Note that `randomMove` has a different distribution, but as we only
need this function for testing and as a placeholder, there is no harm
in that.