finish `fn parseCsa(...)` functionality
We now have a working helper function `parseCsa` in `parse.zig` that
will take any game.csa file (as long as it is for a 'standard' game -
the usual board size and setup, no handicaps for either player) and
translate it to a collection of `model.Move`s.

This will be helpful for debugging and testing the last few issues in
our game logic, as we can now easily check it against real games and
look for any errors. Perhaps in future we could add a CLI option
like `--debug-game-logic-on <game.csa>` for convenience, which would
run such a test instead of launching the game.

Summary of changes in `parse.zig`:

- Optionally, test the moves from `parseCsa` with the `isValid`
  function. We will want to use this to help debug isValid itself, as
  it still has some issues.
- Fix csaPos by translating the x-coord to count left-to-right (what we
  use for our internal representation) instead of right-to-left (normal
  notation used in csa files).
- Statically check array lengths for csaPos and csaSort.
- Add more docstrings
- Rename things for clarity