bugfix when dropping pieces with count=0
For a bit now, there has been a bug when dropping from the users hand
to the board, a piece with count=0 (that is, a piece which the
player doesn't actually have any of in their hand). This situation
triggered an isValid check of the proposed drop which was implicitly
assuming the dropped piece actually had count >0. Eventually this move
would be checked by `model.Board.applyMoveDrop` which would discover it
is actually invalid and panic.

So the fix is easy enough, just add such a check into isValid.